Add margin_short to chart
