+++ web/src/components/MainTab.jsx
@@ -24,7 +24,7 @@ export default function MainTab({
   handleBlur,
   addMin,
   addAll,
-  baseOrderSizeInfoText
+  baseOrderSizeInfoText,
 }) {
   return (
     <TabPane tabId="main">
@@ -54,10 +54,7 @@ export default function MainTab({
       <Row className="u-margin-bottom">
         <Col md="6" sm="12">
           <Label htmlFor="base_order_size">
-            <BotFormTooltip
-              name="base_order_size"
-              text={baseOrderSizeInfoText}
-            >
+            <BotFormTooltip name="base_order_size" text={baseOrderSizeInfoText}>
               Base order size
             </BotFormTooltip>
             <span className="u-required">*</span>
@@ -70,9 +67,7 @@ export default function MainTab({
               onBlur={handleBlur}
               value={bot.base_order_size}
               autoComplete="off"
-              disabled={
-                bot.status === "active" || bot.status === "completed"
-              }
+              disabled={bot.status === "active" || bot.status === "completed"}
             />
             <InputGroupText>{bot.quoteAsset}</InputGroupText>
           </InputGroup>
@@ -140,28 +135,30 @@ export default function MainTab({
         )}
       </Row>
       <Row>
-        <Col md="6">
-          <FormGroup>
-            <BotFormTooltip
-              name="short_sell_price"
-              text="Price at which to stop loss sell and later buy again with short_buy_price (short strategy autoswitch)"
-            >
-              Short Sell Price
-            </BotFormTooltip>
-            <InputGroup>
-              <Input
-                type="number"
+        {bot.strategy === "short" && (
+          <Col md="6">
+            <FormGroup>
+              <BotFormTooltip
                 name="short_sell_price"
-                onBlur={handleBlur}
-                onChange={handleChange}
-                value={bot.short_sell_price}
-                autoComplete="off"
-                step="0.00000001"
-              />
-              <InputGroupText>{bot.quoteAsset}</InputGroupText>
-            </InputGroup>
-          </FormGroup>
-        </Col>
+                text="Price at which to stop loss sell and later buy again with short_buy_price (short strategy autoswitch)"
+              >
+                Short Sell Price
+              </BotFormTooltip>
+              <InputGroup>
+                <Input
+                  type="number"
+                  name="short_sell_price"
+                  onBlur={handleBlur}
+                  onChange={handleChange}
+                  value={bot.short_sell_price}
+                  autoComplete="off"
+                  step="0.00000001"
+                />
+                <InputGroupText>{bot.quoteAsset}</InputGroupText>
+              </InputGroup>
+            </FormGroup>
+          </Col>
+        )}
         <Col md="6" sm="12">
           <FormGroup>
             <BotFormTooltip
@@ -194,31 +191,34 @@ export default function MainTab({
             >
               <option value="long">Long</option>
               <option value="short">Short</option>
+              <option value="margin_short">Margin short</option>
             </Input>
           </FormGroup>
         </Col>
-        <Col md="6" sm="12">
-          <FormGroup>
-            <BotFormTooltip
-              name="short_buy_price"
-              text="Price at which to execute base order"
-            >
-              Short Buy Price
-            </BotFormTooltip>
-            <InputGroup>
-              <Input
-                type="number"
+        {bot.strategy === "short" && (
+          <Col md="6" sm="12">
+            <FormGroup>
+              <BotFormTooltip
                 name="short_buy_price"
-                onChange={handleChange}
-                onBlur={handleBlur}
-                value={bot.short_buy_price}
-                autoComplete="off"
-                step="0.00000001"
-              />
-              <InputGroupText>{bot.quoteAsset}</InputGroupText>
-            </InputGroup>
-          </FormGroup>
-        </Col>
+                text="Price at which to execute base order"
+              >
+                Short Buy Price
+              </BotFormTooltip>
+              <InputGroup>
+                <Input
+                  type="number"
+                  name="short_buy_price"
+                  onChange={handleChange}
+                  onBlur={handleBlur}
+                  value={bot.short_buy_price}
+                  autoComplete="off"
+                  step="0.00000001"
+                />
+                <InputGroupText>{bot.quoteAsset}</InputGroupText>
+              </InputGroup>
+            </FormGroup>
+          </Col>
+        )}
       </Row>
     </TabPane>
   );
